+++ src/utils/jms.js
@@ -131,6 +131,10 @@ export function getRouteViewRequirePerms(route) {
 }
 
 export function hasRouteViewPerm(route) {
+  console.log('Route: ', route)
+  if (route.name) {
+    return hasPermission(route.meta.permissions)
+  }
   const viewName = route.path.split('/')[1]
   const perms = getViewRequirePerms(viewName)
   return hasPermission(perms)
